board/requirements.txt
micropython-esp32-stubs==1.21.0.post4
micropython-stdlib-stubs==1.1.2
pillow==10.2.0
